made database still need to refractor bean to work

+++ src/main/java/com/gcu/business/OrdersBusinessServiceInterface.java
@@ -11,7 +11,7 @@ public interface OrdersBusinessServiceInterface {
 	public void destroy();
 	public OrderModel getById(int id);
 	public List<OrderModel> searchOrders(String searchTerm);
-	public int addOne(OrderModel newOrder);
+	public long addOne(OrderModel newOrder);
 	public boolean deleteOne(long id);
 	public OrderModel updateOne(long idToUpdate, OrderModel updateOrder);
 }
